+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](http://semver.org).
 - Fixed - Wrong argument sometimes reported by "Unknown argument" error.
 - Added - Multivalue options (one option taking multiple arguments).
 - Fixed - opt.defaultDesc("") doesn't suppress a flag value's (default) clause.
+- Added - opt.nameDesc() for overriding first column in options list.
 
 ## dimcli 7.2.0 (2024-02-22)
 - Added - More cli.to\*Cmdline() and cli.to\*CmdlineL() overloads